Add files via upload
The Student Project Management System is a database-driven application designed to manage the allocation and tracking of student projects in an academic setting. It handles the relationships between students, supervisors, projects, and tasks. The system allows for:

Assigning students (limited to 2nd and 3rd year) to various projects.
-> Assigning supervisors to oversee specific projects.
->Tracking tasks associated with each project, including due dates and completion status.
->Monitoring the number of students involved in each project.
->Viewing tasks and supervisors linked to individual projects.

This system ensures better organization and monitoring of academic projects, helping both students and supervisors track progress and responsibilities efficiently.